Make counter private
