+++ packages/aws-cdk/lib/cxapp/cloud-executable.ts
@@ -1,9 +1,9 @@
 import type * as cxapi from '@aws-cdk/cloud-assembly-api';
-import { countAssemblyResults, ToolkitError } from '@aws-cdk/toolkit-lib';
+import { ToolkitError } from '@aws-cdk/toolkit-lib';
 import { CloudAssembly } from './cloud-assembly';
 import type { ICloudAssemblySource, IReadableCloudAssembly } from '../../lib/api';
 import type { IoHelper } from '../../lib/api-private';
-import { BorrowedAssembly } from '../../lib/api-private';
+import { BorrowedAssembly, countAssemblyResults } from '../../lib/api-private';
 import type { SdkProvider } from '../api/aws-auth';
 import { GLOBAL_PLUGIN_HOST } from '../cli/singleton-plugin-host';
 import { cdkCliErrorName } from '../cli/telemetry/error';
